Seperate drawable and updateable flag from Country + Use UpdateableFlag
