+++ osu.Game/Overlays/BeatmapSet/Scores/ScoreTable.cs
@@ -13,7 +13,7 @@ using osu.Game.Graphics.Sprites;
 using osu.Game.Online.Leaderboards;
 using osu.Game.Rulesets.UI;
 using osu.Game.Scoring;
-using osu.Game.Users;
+using osu.Game.Users.Drawables;
 using osuTK;
 using osuTK.Graphics;
 
@@ -135,7 +135,7 @@ namespace osu.Game.Overlays.BeatmapSet.Scores
                     Spacing = new Vector2(5, 0),
                     Children = new Drawable[]
                     {
-                        new DrawableFlag(score.User.Country) { Size = new Vector2(20, 13) },
+                        new UpdateableFlag(score.User.Country) { Size = new Vector2(20, 13) },
                         username
                     }
                 },
